cleaning up test, moving to auth
